separate main code #234
