+++ src/lib/components/backgrounds/AnimatedBackground.svelte
